+++ ATTINYCellModuleMigrate/ATTINYCellModule/ATTINYCellModule/util/eeprom.h
@@ -5,7 +5,7 @@
 
 typedef enum configs { VOLT_CALIB } configs_t;
 
-uint16_t get_config(configs_t config);
-void set_config(configs_t config, uint16_t value);
+float get_config(configs_t config);
+void set_config(configs_t config, float value);
 
 #endif
